leger modif

+++ app/src/main/java/evaluateur/InterfaceEvaluateur.java
@@ -1,4 +1,4 @@
 package evaluateur;
 public interface InterfaceEvaluateur {
-    int donneLeScore();
+    double donneLeScore();
 }
